Move queryOptions to api folder
