+++ frontend/src/features/requests/RequestDatesDialog.tsx
@@ -15,7 +15,7 @@ import { PickerValue } from '@mui/x-date-pickers/internals';
 import dayjs, { Dayjs } from 'dayjs';
 import { useCallback, useEffect, useState } from 'react';
 import { TARequestDetail, TARequestDetailMutation } from '../../api/dashboard/types';
-import { useRequestMutation } from '../../utils/queryOptions';
+import { useRequestMutation } from '../../api/queryOptions';
 import { useIdentityContext } from '../identity/IdentityContext';
 import { useToastContext } from '../toasts/ToastContext';
 import { ToastMessage } from '../toasts/ToastMessage';
